+++ tests/wl_test_lang_examples.py
@@ -1,5 +1,5 @@
 # ----------------------------------------------------------------------
-# Wordless: Tests - Language Examples
+# Wordless: Tests - Language examples
 # Copyright (C) 2018-2023  Ye Lei (叶磊)
 #
 # This program is free software: you can redistribute it and/or modify