+++ tests/wl_tests_checks/test_checks_work_area.py
@@ -1,5 +1,5 @@
 # ----------------------------------------------------------------------
-# Wordless: Tests - Checks - Work Area
+# Wordless: Tests - Checks - Work area
 # Copyright (C) 2018-2023  Ye Lei (叶磊)
 #
 # This program is free software: you can redistribute it and/or modify