+++ tests/wl_tests_nlp/test_dependency_parsing.py
@@ -1,5 +1,5 @@
 # ----------------------------------------------------------------------
-# Wordless: Tests - NLP - Dependency Parsing
+# Wordless: Tests - NLP - Dependency parsing
 # Copyright (C) 2018-2023  Ye Lei (叶磊)
 #
 # This program is free software: you can redistribute it and/or modify