+++ tests/wl_tests_work_area/test_colligation_extractor.py
@@ -1,5 +1,5 @@
 # ----------------------------------------------------------------------
-# Wordless: Tests - Work Area - Colligation Extractor
+# Wordless: Tests - Work area - Colligation Extractor
 # Copyright (C) 2018-2023  Ye Lei (叶磊)
 #
 # This program is free software: you can redistribute it and/or modify