+++ utils/wl_trs_generate_ts_files.py
@@ -1,5 +1,5 @@
 # ----------------------------------------------------------------------
-# Utilities: Translations - Generate TS Files
+# Utilities: Translations - Generate TS files
 # Copyright (C) 2018-2023  Ye Lei (叶磊)
 #
 # This program is free software: you can redistribute it and/or modify