+++ wordless/wl_nlp/wl_sentence_tokenization.py
@@ -1,5 +1,5 @@
 # ----------------------------------------------------------------------
-# Wordless: NLP - Sentence Tokenization
+# Wordless: NLP - Sentence tokenization
 # Copyright (C) 2018-2023  Ye Lei (叶磊)
 #
 # This program is free software: you can redistribute it and/or modify