+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 ## [3.3.0](https://github.com/BLKSerene/Wordless/releases/tag/3.3.0) - ??/??/2023
 ### 🎉 New Features
 - Settings: Add Settings - Measures - Readability - Automated Readability Index / Bormuth's Grade Placement / Coleman's Readability Formula / Dale-Chall Readability Formula / Danielson-Bryan's Readability Formula / Flesch Reading Ease / Flesch Reading Ease (Farr-Jenkins-Paterson) / Gunning Fog Index / Spache Grade Level / Tränkle & Bailer's Readability Formula
+- Settings: Add Settings - Sentiment Analysis
+- Utils: Add Dostoevsky's Russian sentiment analyzer
 - Utils: Add khmer-nltk's Khmer sentence tokenizer, word tokenizer, and part-of-speech tagger
 - Utils: Add PyThaiNLP's perceptron part-of-speech tagger (Blackboard)
 - Utils: Add spaCy's Korean sentence recognizer, word tokenizer, part-of-speech tagger, lemmatizer, and dependency parser
@@ -42,6 +44,7 @@
 - Work Area: Remove Profiler - Fernández Huerta's Readability Score / Szigriszt's Perspicuity Index
 
 ### ⏫ Dependency Changes
+- Dependencies: Add Dostoevsky
 - Dependencies: Add khmer-nltk
 - Dependencies: Remove TextBlob
 - Dependencies: Upgrade Beautiful Soup to 4.12.2